+++ main.cpp
@@ -41,7 +41,7 @@ public:
 	{
 		AttachLayer(new Layer1());
 		AttachLayer(new Layer2());
-		LimitFPS(CSE_FPS60);
+		LimitFPS(CSE_FPS10);
 	};
 	~App() {};
 };